Configure sanitizers in CMake, fix issues (BlueBrain/CoreNeuron#823)
* Use sanitizer helpers from hpc-coding-conventions.
* Integrate sanitizer-specific configuration into CMake
  configuration so less external magic is needed.
* Add ASan and UBSan jobs to the GitHub Actions CI.
  LSan is not currently enabled because of MPI.
* Add problem matchers to GitHub Actions CI so that
  compiler warnings and sanitizer issues are flagged.
* Drop two redundant macOS GitHub Actions CI jobs.
* Use ccache in GitHub Actions CI.
* Add UBSan exclusion file.
* Do not pass -undefined dynamic_lookup to vanilla Clang.
* Fix signed/unsigned comparison compiler warnings.
* Fix struct/class mismatch compiler warnings.
* stdindex2ptr: fix bugs when nrn_sav_rhs is too far away
  from _data in memory.
* Fix __restrict__ violation in sprintf.

CoreNEURON Repo SHA: BlueBrain/CoreNeuron@6215225fe2814648f55d9457a79d373f72df62c2

+++ src/coreneuron/io/mech_report.cpp
@@ -21,7 +21,7 @@ void write_mech_report() {
     std::vector<long> local_mech_count(n_memb_func, 0);
 
     /// each gid record goes on separate row, only check non-empty threads
-    for (size_t i = 0; i < nrn_nthread; i++) {
+    for (int i = 0; i < nrn_nthread; i++) {
         const auto& nt = nrn_threads[i];
         for (auto* tml = nt.tml; tml; tml = tml->next) {
             const int type = tml->index;
